adding doc

+++ ft_transcendence_db/src/entity/friendships/friendship.service.ts
@@ -8,10 +8,12 @@ export class FriendshipService {
         @Inject('FRIENDSHIP_REPOSITORY')
         private friendshipRepository: Repository<Friendships>,
     ) {}
-    async findAll(): Promise<Friendships[]>
-    {
-        return this.friendshipRepository.find();
-    }
+    /**
+     * get the friend list of the user
+     * 
+     * @param id id of user
+     * @returns the friend list of the user
+     */
     async findFriendOfId(id: number)
     {
         const friendshipsFromFriend = await this.friendshipRepository
@@ -26,7 +28,6 @@ export class FriendshipService {
             .innerJoinAndSelect('friendships.friend', 'friend')
             .where(`friend.id = :id AND statu = 'accepted'`, { id: id })
             .getMany();
-        
         const usersFromFriend = friendshipsFromFriend.map((friendship) => ({
             id: friendship.friend.id, 
             pseudo: friendship.friend.pseudo, 
@@ -39,6 +40,12 @@ export class FriendshipService {
             isConnected: friendship.user.isConnected}));
         return ([...usersFromFriend, ...usersFromUser]);
     }
+    /**
+     * Get the firend request that a user received
+     * 
+     * @param id id of user
+     * @returns the friend requests of the user
+     */
     async findPending(id: number)
     {
         const friensdAsking = await this.friendshipRepository
@@ -55,6 +62,14 @@ export class FriendshipService {
             isConnected: friendship.user.isConnected}));
         return (usersAsking);
     }
+    /**
+     * accept a friendship request
+     * 
+     * @param friend_id id of user accpeting the friendship request
+     * @param user_id id of user that sent the request
+     * @returns result of the request
+     * @throws Error('Friendship not found.')
+     */
     async acceptFriendship(friend_id: number, user_id: number): Promise<Friendships> 
     {
         const friendship = await this.friendshipRepository.findOne({
@@ -68,12 +83,27 @@ export class FriendshipService {
         friendship.statu = 'accepted';
         return this.friendshipRepository.save(friendship);
     }
+    /**
+     * remove a friendship (that can be either accepted or pending)
+     * 
+     * @param friend_id id of one of the friends
+     * @param user_id id of one of the friends
+     * @returns result of request
+     * @throws Error('no such friendship.')
+     */
     async deleteFriendship(friend_id: number, user_id: number)
     {
-        const friendship = await this.friendshipRepository.findOne({
+        let friendship = await this.friendshipRepository.findOne({
             where: { user: { id: user_id }, friend: { id: friend_id } },
-          });
-
+            });
+        if (!friendship)
+        {
+            friendship = await this.friendshipRepository.findOne({
+                where: { user: { id: friend_id }, friend: { id: user_id } },
+                });
+        }
+        if (!friendship)
+            throw new Error('no such friendship.');
         const res = await this.friendshipRepository.delete(friendship.id);
         return (res);
     }
